Moved binary guid extraction to a new file and de-duplicated some code
