+++ common/src/jetbrains/buildServer/symbols/PdbSignatureIndexUtil.java
@@ -49,8 +49,8 @@ class PdbSignatureIndexUtil {
     XmlUtil.saveDocument(new Document(root), outputStream);
   }
 
-  private static String extractGuid(String sign, boolean withDebugType) {
-    if (withDebugType)
+  private static String extractGuid(String sign, boolean cutDebugType) {
+    if (cutDebugType)
       return sign.substring(0, sign.length() - 1).toLowerCase(); //last symbol is PEDebugType
     else
       return sign.toLowerCase();
